+++ src/app/api/friends/add/route.ts
@@ -2,7 +2,7 @@ import {addFriendValidator} from "@/lib/validations/add-friend";
 import {Utils} from "@/lib/utils";
 import myGetServerSession from "@/lib/myGetServerSession";
 import {pusherServer} from "@/lib/pusher";
-import fetchRedis from "@/app/helpers/redis";
+import fetchRedis from "@/helpers/redis";
 import {db} from "@/lib/db";
 
 interface errorProps{
